+++ Angular2Emulator/TypeScript/Ng2Emulation/Core/Bootstrap.ts
@@ -2,7 +2,12 @@
 import ElementEvents from "../Events/ElementEvents";
 import httpIntercept from "../Templates/HttpInterceptor";
 import decorateInterpolate from "./InterpolateDecorator"
+import decorateParse from "./ParseDecorator"
 import {NgProperty} from "../Directives/NgProperty";
+import {NgEventBinding} from "../Directives/NgEventBinding";
+import {NgPropertyBinding} from "../Directives/NgPropertyBinding";
+import decorateController from "./ControllerDecorator";
+import {initLifeCycleHooks} from "./LifeCycle/LifeCycleHooks";
 
 /**
  * Class to store bootstrap information.
@@ -20,17 +25,25 @@ export class BootStrapper {
         // Interceptor for templates.
         httpIntercept(Angular1Wrapper.app);
 		// Decorate $interpolate to redirect scope to $$cmp
-	    decorateInterpolate(Angular1Wrapper.app);
+	    //decorateInterpolate(Angular1Wrapper.app);
+        decorateParse(Angular1Wrapper.app);
+        decorateController(Angular1Wrapper.app);
 
         // Register built-in directives
         // register element events directives.
-        ElementEvents.resolve().forEach(directive => Angular1Wrapper.registerDirective(directive));
+        //ElementEvents.resolve().forEach(directive => Angular1Wrapper.registerDirective(directive));
+
+		// register Event Binding directive
+		Angular1Wrapper.registerDirective(NgEventBinding);
+		// register Property Binding directive
+		Angular1Wrapper.registerDirective(NgPropertyBinding);
         // register ngProperty
         Angular1Wrapper.registerDirective(NgProperty);
         
         Angular1Wrapper.registerServices(BootStrapper.services);
         Angular1Wrapper.registerComponent(component);
 
+        initLifeCycleHooks(Angular1Wrapper.app);
         Angular1Wrapper.bootstrap();
     }
 }