+++ Angular2Emulator/TypeScript/Ng2Emulation/Decorators/Directive.ts
@@ -1,5 +1,8 @@
 ﻿export interface IDirectiveMetadata {
     selector?: string;
+	priority?: number;
+	outputs?: string[];
+	inputs?: string[];
 }
 
 /**