+++ Angular2Emulator/TypeScript/Ng2Emulation/Ng2Emulation.ts
@@ -1,4 +1,12 @@
-﻿export * from "./Decorators/Component"
-export * from "./Decorators/Directive"
-export * from "./Decorators/Inject"
-export * from "./Core/Bootstrap"
+﻿/// <reference path="../../scripts/typings/es6-promise/es6-promise.d.ts" />
+
+export * from "./Decorators/Component";
+export * from "./Decorators/Directive";
+export * from "./Decorators/Inject";
+export * from "./Decorators/Injectable";
+export * from "./Events/EventEmitter";
+export * from "./Core/Bootstrap";
+export * from "./Decorators/Output";
+export * from "./Decorators/Input";
+export {OnChanges, OnInit, OnDestroy} from "./Core/LifeCycle/LifeCycleHooks";
+