+++ Angular2Emulator/TypeScript/Ng2Emulation/Templates/Parser.ts
@@ -1,9 +1,23 @@
 ﻿import {ParserRule} from "./ParserRule";
 import NgContentRule from "./rules/NgContentRule"
+import NgModelRule from "./rules/NgModelRule"
 import NgPropertyRule from "./rules/NgPropertyRule"
+import EventBindingRule from "./rules/EventBindingRule"
+import PropertyBindingRule from "./rules/PropertyBindingRule"
+import TwoWayBindingRule from "./rules/TwoWayBindingRule"
+import NgForRule from "./rules/NgForRule"
+import NgIfRule from "./rules/NgIfRule"
 
 export default class Parser {
-    private static rules: ParserRule[] = [new NgContentRule(), new NgPropertyRule()];
+    private static rules: ParserRule[] = [
+		new NgModelRule(), new NgContentRule(), new NgPropertyRule(),
+		new TwoWayBindingRule(), new PropertyBindingRule(), new EventBindingRule(),
+		new NgForRule(), new NgIfRule()
+	];
+    //private static rules: ParserRule[] = [
+	//	new NgModelRule(), new NgContentRule(), new NgPropertyRule(),
+	//	new TwoWayBindingRule(), new PropertyBindingRule(), new EventBindingRule()
+	//];
 
     static addRule(rule: ParserRule) {
         this.rules.push(rule);