+++ CHANGELOG.md
@@ -9,6 +9,26 @@ instead of
 ````html
 <div>{{component.componentProperty}}</div>
 ````
+* `(event-binding)` search in component for custom (outputs) events (EventEmitter), if not found try to emit element event.
+* `[property-binding]` search in component for custom (inputs) properties, if not found assign value to element property.
+* `@Input` and `@Output` add property to `$componentMetadata.inputs` and `$componentMetadata.outputs`
+* Cannonical syntax is accepted (`bind-propertyName`, `on-eventName`, `var-localVariableName`)
+* Attribute syntax `[attr.colspan]="expression"`.
+* Class syntax `[class.red]="amount < 0"`.
+* Style syntax `[style.color] = "isSpecial ? 'red' : 'green'"`. Or with units: `[style.fontSize.em]="isSpecial ? 3 : 1"`
+* NgModelRule for template parser, replace [(ngModel)]="some.thing" by ng-model="$$cmp.some.thing"
+* *ngFor working. Replace *ngFor by ng-repeat + ng-init: 
+````html
+<div *ngFor="#item of itemList, #i = index"></div>
+````
+change by:
+````html
+<div ng-repeat="item in itemList" ng-init="i = $index"></div>
+````
+* styles added to component metadata and load it in bootstrap (added globally) 
+* Add `providers` to @Component metadata.
+* LifeCycleHooks - OnChanges, OnInit and OnDestroy
+
 
 ## Version 0.2 ALPHA
 * **breaking** controllerAs is $$cmp instead of $$vm
@@ -20,7 +40,7 @@ instead of
 ````
 * NgPropertyRule to replace #something by ng-property="something", to access to html like angular2 you can do:
 ````html
- <input type="text" #mybutton/> {{$cmp.myinput.value}}
+ <input type="text" #myinput/> {{$cmp.myinput.value}}
 ````
 
 ## Version 0.1 ALPHA